// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TemplateStatus = {
  ACTIVE: 'ACTIVE',
  PLANNED: 'PLANNED',
  ARCHIVED: 'ARCHIVED'
} as const

export type TemplateStatus = (typeof TemplateStatus)[keyof typeof TemplateStatus]


export const WebsiteStatus = {
  DRAFT: 'DRAFT',
  PREVIEW: 'PREVIEW',
  PUBLISHED: 'PUBLISHED',
  EXPIRED: 'EXPIRED',
  ARCHIVED: 'ARCHIVED'
} as const

export type WebsiteStatus = (typeof WebsiteStatus)[keyof typeof WebsiteStatus]


export const MediaType = {
  IMAGE: 'IMAGE',
  VIDEO: 'VIDEO',
  AUDIO: 'AUDIO'
} as const

export type MediaType = (typeof MediaType)[keyof typeof MediaType]


export const PaymentStatus = {
  UNPAID: 'UNPAID',
  PARTIAL: 'PARTIAL',
  PAID: 'PAID',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const OrderStatus = {
  NEW: 'NEW',
  WAITING_PAYMENT: 'WAITING_PAYMENT',
  PAID: 'PAID',
  PROCESSING: 'PROCESSING',
  REVISION: 'REVISION',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]